pyproject.toml
[project]
name = "fastapi-starter-pack"
version = "0.0.1"
description = ""
authors = [
    { name = "Kamyab Vijapura", email = "kamyab.vijapura@webelight.co.in" }
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.128.0",
    "sqlalchemy[postgresql-asyncpg]>=2.0.30",
    "alembic>=1.13.1",
    "fastapi-pagination>=0.15.0",
    "fastapi-redis-cache>=0.2.5",
    "pydantic-settings>=2.3.3",
    "pyjwt>=2.6.0",
    "passlib>=1.7.4",
    "httpx>=0.28.0",
    "apscheduler>=3.10.1",
    "uvicorn[standard]>=0.40.0",
    "pydantic-extra-types>=2.8.2",
    "fastapi-limiter==0.1.6",
    "asyncmy>=0.2.9",
    "colorama>=0.4.6",
    "cryptography>=46.0.0",
    "bcrypt>=5.0.0",
    "pytz>=2025.2",
    "python-multipart>=0.0.21",
    "email-validator>=2.3.0",
    "rich>=14.2.0",
    "typer>=0.21.1",
    "psutil>=7.2.1",
    "sentry-sdk>=2.50.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "aiosqlite>=0.20.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
